refactor(tests): watcher.rs (3096 lines) → grouped watcher/ binary
Same #191 playbook as the reducer split: tests/watcher.rs becomes the
grouped binary tests/watcher/main.rs with six mechanism-scoped submodules:

- tailing.rs     cursor mechanics: append-tail emit, partial trailing line,
                 truncation reset, non-UTF-8 skip (4 tests)
- first_sight.rs the first-sight gate: stale/recent/ended/oversized seeds,
                 probe bypass, cwd extraction, id/label derivers, subagent
                 parent links (15 tests + custom_label deriver)
- liveness.rs    proof-of-life emission, negative vouch, instant exit (pid
                 death), probe-failure no-ops (9 tests +
                 vouch_snapshot_with_pid / admitted_with_mutable_probe /
                 assert_no_session_end_within fixtures)
- unclaim.rs     child-end un-claim (#246/#250): turn-N+1 re-register +
                 in-flight multi-turn revival (2 tests)
- sources.rs     Source::run glue: codex / antigravity / claude-code
                 bind+spawn (3 tests)
- attach.rs      the mid-attach scenario suite — attach shows exactly the
                 live set (4 tests)

Moves only — every fn item verified byte-identical against HEAD; the
cross-module poll-seam harness (fast_watch, cc_watcher, vouch_snapshot,
write_lines, backdate, cc_*_line builders) lives in main.rs; the
liveness-only fixtures move with their suite. Only mod decls and use lines
changed. nextest list: 634 → 634 (watcher binary 37 → 37); full workspace
1432/1432 green incl. x86_64-pc-windows-msvc clippy.
